+++ product_outline_extraction.py
@@ -210,7 +210,7 @@ def product_outline_extraction_by_mask_multiple_product_types(args, grounding_mo
         image_array = np.asarray(img)
 
         #white_array = np.ones_like(image_array) * args.hed_value
-        white_array = np.ones((image_dim, image_dim, 3)) * args.hed_value
+        white_array = np.ones((image_dim, image_dim, 3), dtype=np.uint8) * args.hed_value
         white_array = white_array * mask_all
         white_array = white_array * mask
 
@@ -779,7 +779,7 @@ def image_outline_re_extraction_by_mask_multiple_product_types(grounding_model, 
     image_array = np.asarray(img)
 
     #white_array = np.ones_like(image_array) * args.hed_value
-    white_array = np.ones((image_dim, image_dim, 3)) * args.hed_value
+    white_array = np.ones((image_dim, image_dim, 3), dtype=np.uint8) * args.hed_value
     white_array = white_array * mask_all
     white_array = white_array * mask
